Implement a Redux data store for the TodoList
The app now has a special page dedicated to a Redux version of the
TodoList. You'll notice that if you switch the page while on the regular
version of the TodoList all of the state is lost. This is _not_ the case
with the Redux version!

The `IDEAS.md` file was also updated to reflect some ideas that
@aaronbatilo had for the project.