Rename the healthcheck endpoint in the server and clean HealthCheck
The healthcheck endpoint is now updated in the server to reflect the
accurate name the enpoint should be. It is no longer 'testing'. The
HealthCheck component has also been 'cleaned' and updated to a nicer
async callback function and sets appropriate styling based on the
message that comes back from the server or failure to reach the server.