Remove the RootReducerRedux and modify the Redux state for database
In preparation for starting to send and receive data to and from the
server I removed the extraneous root reducer code as well as updated the
components respectively.

The code for send/receive is in the `ServerRedux.js` file but it has not
been hooked up to the actual reducer code just yet. I still need to
figure out how exactly I want to structure the todos to MongoDB.