Send and receive data to and from MongoDB
The frontend now both sends and receives information to and from the
server to access a MongoDB instance.

In the `indexRedux.js` file we now try to grab from the database the
list of todos to send to the `TodoListRedux` component. Otherwise, we
just set the state of the list to an empty one. This should be more
robust. We _should_ check and see if we really can get data back from
the server, whether the server is down, etc.

In the reducer function, as we set the state of the frontend components
we also call out to asyncronous functions that will try to send data
back to the server to update the database. This should _also_ be more
robust. It was suggested to use a "thunk" pattern where we can try to
update the server within the action call and then return a success or
failure and update the state of the frontend based on that.

I also added flake8 as a linter for the backend python files.

There is still a lot more work to do, including thinking about how to
implement tests, CI/CD integration, as well as deployment (probably to
Heroku).