Send actions through Redux middleware thunks based on server response
I've added the `redux-thunk` middleware to Redux so now we base how the
state of the application updates based on the how the server responds to
requests sent from the client. If the server does not respond, the
client does not change.

This may not be the exact behavior we want, we
could look at trying to display a "saved" message to the user or
something along those lines. However, this is a good interim solution so
the state of the server and client are now lock-stepped so their states
can never differ.

MongoDB now also stores the state of the 'checkall' checkbox so the
issue we were having regarding its state not being reflected after a
refresh is now fixed.